updated the file sturcture
